fix: Bad Requests when targetIDs was empty.
If the list of target IDs is empty, we send no add list and no remove
list for a link. The AIPE does not like that and return a 400 - Bad
Request in such a case.

However, if we are supposed to remove 0 target IDs, or add 0 target IDs,
or change no target IDs, why talk to the AIPE at all?