change: only return properties from read the user cares about.
Prior to this, the read always returned the entire object. However,
sometimes users want to maintain only a subset of object properties via
terraform, and manage other parts of the object using the Workflows and
Forms of the AI Process Engine.

The prior behavior caused this to constantly result in diffs, which
however were not updated at all, so they stuck around forever.

With this change, only the properties the user adds are returned and
considered for terraform diffs.